Small improvements to integration tests run-local.sh

+++ integration-tests/run-local.sh
@@ -15,17 +15,37 @@ set -ex
 : ${SUFFIX:=$(git rev-parse HEAD)}
 : ${COMPLETER_DOCKER_IMAGE:=fnproject/completer}
 
+# ----------------------------------------------------------------------
+# Check for prerequisites
+# ----------------------------------------------------------------------
+
+if [[ `echo nooo | sed -e "s/nooo/ok/"` != "ok" ]]; then
+    echo "You need GNU sed installed, because this script uses 'sed -e'."
+    echo "If you are on Mac, run 'brew install gnu-sed --with-default-names'."
+    exit 2
+fi
+
+if [[ `echo grepme | grep -e "grepme"` != "grepme" ]]; then
+    echo "You need GNU grep installed, because this script uses 'grep -e'."
+    echo "If you are on Mac, run 'brew install gnu-grep --with-default-names'."
+    exit 2
+fi
 
 # ----------------------------------------------------------------------
 # Stand up a local staging maven directory, if needed
 # ----------------------------------------------------------------------
 
 if [[ -n "$REPOSITORY_LOCATION" ]]; then
-    export MAVEN_REPOSITORY_LOCATION="http://172.17.0.1:18080"
+    if [[ -n "$LOCALHOST_ACTUAL_IP" ]]; then
+        echo "Using $LOCALHOST_ACTUAL_IP for the staging Maven repo."
+        export MAVEN_REPOSITORY_LOCATION="http://$LOCALHOST_ACTUAL_IP:18080"
+    else
+        echo "No LOCALHOST_ACTUAL_IP set, assuming we're running on CircleCI."
+        export MAVEN_REPOSITORY_LOCATION="http://172.17.0.1:18080"
+    fi
     cd "$REPOSITORY_LOCATION" && python -mSimpleHTTPServer 18080 1>>/tmp/http-logs 2>&1 &
     defer kill -9 "$!"
 fi
-
 
 # ----------------------------------------------------------------------
 # Stand up the functions platform
